pyproject.toml
[project]
name = "hyperliquid-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "ccxt>=4.4.0",
    "websockets>=13.0",
    "orjson>=3.10.0",
    "coincurve>=20.0",
    "pandas>=2.2.0",
    "pandas-ta>=0.3.14b1",
    "numpy>=1.26.0",
    "anthropic>=0.40.0",
    "aiosqlite>=0.20.0",
    "python-telegram-bot>=21.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
    "uvloop>=0.21.0",
    "tenacity>=9.0.0",
    "schedule>=1.2.0",
]

[project.scripts]
trading-bot = "main:main_entry"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
